python/batch/pyproject.toml
[tool.poetry]
name = "deutschland-ticket"
version = "0.1.0"
description = ""
authors = ["Vikram <vikramsingh001@gmail.com>"]
readme = "README.md"
packages = [{include = "deutschland_ticket"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<3.11"
pydantic = "^1.10.7"
requests = "^2.30.0"
beautifulsoup4 = "^4.12.2"
python-dotenv = "^1.0.0"
langchain = "^0.0.188"
text-generation = "^0.5.2"
openai = "^0.27.7"
tiktoken = "^0.4.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.3.0"
black = "^23.3.0"
types-requests = "^2.30.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
